requirements.txt
# ktb-pod dependencies
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0
pytz>=2023.3
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
piexif>=1.1.3

# ktb-upload dependencies (ktb-upload/requirements.txt cũ)
paramiko>=3.0.0
python-dotenv>=1.0.0
